# Improve parameters performances