On `france`:
-  `nosetests tests/test_basics.py` drops from `35s` to `25s`
-  [This file](https://github.com/openfisca/tutorial/blob/master/python/scripts/generate_situation_examples/try_situation_examples.py) drops from `5.26s` to `3.55s`

Interesting lessons:
- Performances waste does not necessarily happen where you think it does
- Flexibles user-friendly interfaces can ruin performances if called under the hood