@fpagnoux 2018-04-13
> How did you discover that `'amount'` in bracket is faster then `hasattr(bracket, 'amount')`

By profiling and discovering that weirdly a lot of time was spent creating `ParameterNotFound` errors, which didn't think to make sense. 

I turned out that each time, `hasattr` calls `__getattr__`, which raise a `ParameterNotFound` if the attributes doesn't exist, and catches it under the hood (as it's a subclass of AttributeError). That's more work than just checking if the item is in a list. 

I also found [resources](https://hynek.me/articles/hasattr/) indicating that `hasattr` is generally not a good option.