@fpagnoux 2018-04-13
For the record:
- I tried to optimize [this function](https://github.com/openfisca/openfisca-core/blob/master/openfisca_core/parameters.py#L165-L168) using [binary search](https://en.wikipedia.org/wiki/Binary_search_algorithm), but it was counter productive. Probably because parameters usually don't have a lot of historical values, or because we tend to look for the most recent ones.

- I think there is a time to be saved by lazy evaluating parameters: each time someone needs a parameter for a specific date, the whole tree is evaluated for this time, which seems overkill.